+++ NEWS.md
@@ -1,4 +1,4 @@
-# **kenpomR 0.2.0-2**
+# **kenpomR 0.2.0-3**
 
 - Added support for ESPN's play-by-play endpoints with the addition of the following functions:
 - ```kenpomR::cbb_espn_game_all()``` - a convenience wrapper function around the following three functions (returns the results as a list of three data frames)
@@ -8,6 +8,8 @@
 - ```kenpomR::cbb_espn_teams()``` (bumps to v0.2.1)
 - ```kenpomR::cbb_espn_conferences()``` (bumps to v0.2.1)
 - ```kenpomR::cbb_espn_scoreboard()``` (bumps to v0.2.2)
+- ```kenpomR::cbb_ncaa_NET_rankings()``` (bumps to v0.2.3)
+- ```kenpomR::cbb_rankings()``` (bumps to v0.2.3)
 
 # **kenpomR 0.1.0**